Move rekey tests to their own file
